две кнопки оплаты. далее планируется сделать вывод нужной кнопки исходя из данных в чекбоксе.
